+++ src/main/java/com/example/modiraa/client/NaverApiClient.java
@@ -1,10 +1,10 @@
 package com.example.modiraa.client;
 
+import com.example.modiraa.dto.request.oauth.OAuthLoginParams;
+import com.example.modiraa.dto.request.oauth.OAuthProvider;
 import com.example.modiraa.dto.response.NaverInfoResponse;
 import com.example.modiraa.dto.response.NaverTokens;
 import com.example.modiraa.dto.response.OAuthInfoResponse;
-import com.example.modiraa.model.oauth.OAuthLoginParams;
-import com.example.modiraa.model.oauth.OAuthProvider;
 import com.example.modiraa.properties.NaverProperties;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpEntity;
