+++ src/main/java/com/example/modiraa/dto/request/oauth/KakaoLoginParams.java
@@ -1,4 +1,4 @@
-package com.example.modiraa.model.oauth;
+package com.example.modiraa.dto.request.oauth;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
