+++ src/main/java/com/example/modiraa/dto/response/OAuthInfoResponse.java
@@ -1,6 +1,6 @@
 package com.example.modiraa.dto.response;
 
-import com.example.modiraa.model.oauth.OAuthProvider;
+import com.example.modiraa.dto.request.oauth.OAuthProvider;
 
 public interface OAuthInfoResponse {
     String getId();
