+++ src/main/java/com/example/modiraa/service/OAuthLoginService.java
@@ -5,11 +5,11 @@ import com.example.modiraa.config.jwt.AuthTokens;
 import com.example.modiraa.config.jwt.AuthTokensGenerator;
 import com.example.modiraa.config.jwt.JwtProperties;
 import com.example.modiraa.dto.request.AdditionalInfoRequest;
+import com.example.modiraa.dto.request.oauth.OAuthLoginParams;
+import com.example.modiraa.dto.request.oauth.OAuthProvider;
 import com.example.modiraa.dto.response.OAuthInfoResponse;
 import com.example.modiraa.dto.response.SocialResponse;
 import com.example.modiraa.model.Member;
-import com.example.modiraa.model.oauth.OAuthLoginParams;
-import com.example.modiraa.model.oauth.OAuthProvider;
 import com.example.modiraa.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
